requirements.txt
deep_sort_realtime==1.3.2
Flask==3.0.3
opencv_contrib_python==4.6.0.66
opencv_python==4.9.0.80
opencv_python_headless==4.6.0.66
PyYAML==6.0.2
PyYAML==6.0.2
ultralytics==8.3.23
